pyproject.toml
[project]
name = "wecare-ai-tickets"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
]
